{{ cookiecutter.project_name.lower().replace(' ', '_').replace('-', '_') }}/requirements.txt
click
Sphinx
coverage
python-dotenv
poutyne
mlflow
numpy
notif
yapf
pytest
hydra-core
hydra-colorlog
torch
torchvision